requirements.txt
pyyaml>=5.4.0,<=5.5
RPi.GPIO>=0.7.0,<=0.8
